services/shopify_scraper/requirements.txt
Flask==0.11.1
python-dotenv==0.15.0
APScheduler==3.6.3
Flask-APScheduler==1.11.0
requests==2.25.1
tqdm==4.55.2
wget==3.2
boto3==1.17.2
python-jose-cryptodome==1.3.2
flask-cors==3.0.2
six==1.13.0
google-api-python-client==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4